requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
arabic-reshaper==3.0.0
asgiref==3.8.1; python_version >= '3.8'
asn1crypto==1.5.1
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
cffi==1.17.0; platform_python_implementation != 'PyPy'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; platform_system == 'Windows'
cron-descriptor==1.4.3
cryptography==43.0.0; python_version >= '3.7'
cssselect2==0.7.0; python_version >= '3.7'
django==3.2.25; python_version >= '3.6'
django-background-tasks==1.2.5
django-bootstrap5-form==5.0.1
django-celery-beat==2.6.0
django-compat==1.0.15
django-paypal==2.1; python_version >= '3.7'
django-phonenumber-field[phonenumberslite]==8.0.0; python_version >= '3.8'
django-timezone-field==7.0; python_version >= '3.8' and python_version < '4.0'
django-widget-tweaks==1.5.0; python_version >= '3.8'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.7; python_version >= '3.5'
kombu==5.4.0; python_version >= '3.8'
lxml==5.3.0; python_version >= '3.6'
oscrypto==1.3.0
phonenumberslite==8.13.43
pillow==10.4.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psycopg2==2.9.9; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pyhanko==0.25.1; python_version >= '3.8'
pyhanko-certvalidator==0.26.3; python_version >= '3.7'
pypdf==4.3.1; python_version >= '3.6'
pypng==0.20220715.0
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pyyaml==6.0.2; python_version >= '3.8'
qrcode==7.4.2; python_version >= '3.7'
redis==5.0.8; python_version >= '3.7'
reportlab==4.2.2; python_version >= '3.7' and python_version < '4'
requests==2.32.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.1; python_version >= '3.8'
svglib==1.5.1; python_version >= '3.7'
tinycss2==1.3.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
tzlocal==5.2; python_version >= '3.8'
uritools==4.0.3; python_version >= '3.7'
urllib3==2.2.2; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16; python_version >= '3.8'
